+++ backend/src/test/java/moadong/util/annotations/IntegrationTest.java
@@ -1,10 +1,7 @@
 package moadong.util.annotations;
 
-import moadong.global.util.TestTypeConstants;
 import org.junit.jupiter.api.Tag;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.ContextConfiguration;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
